% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzhCovid19Docs.R
\docType{data}
\name{covid_docs}
\alias{covid_docs}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{release}{}
\item{category}{}
\item{title}{}
\item{link}{}
\item{doc_date}{}
\item{doc_type}{}
\item{page}{}
\item{text}{}
\item{ocr}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{nzhCovid19Docs}
}
