% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzhCovid19Docs.R
\docType{package}
\name{nzhCovid19Docs}
\alias{nzhCovid19Docs}
\alias{nzhCovid19Docs-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
nzhCovid19Docs
A data package for nzhCovid19Docs.

Use \code{data(package='nzhCovid19Docs')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{covid_docs}
}
