% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandemicpapers.R
\docType{package}
\name{pandemicpapers}
\alias{pandemicpapers}
\alias{pandemicpapers-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
pandemicpapers
A data package for pandemicpapers.

Use \code{data(package='pandemicpapers')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{covid_docs}
}
